Avoid use of spinlock when possible

+++ src/services/pcn-policer/src/Contract.cpp
@@ -121,7 +121,7 @@ void Contract::updateData(ContractUpdateDataInputJsonObject input) {
 
 void Contract::updateDataplane() {
   struct bucket bucket = {
-    .tokens = burst_limit_ * 1000,
+    .tokens = (int64_t)burst_limit_ * 1000,
     .refill_rate = rate_limit_,
     .capacity = burst_limit_ * 1000,
     .last_refill = 0
